requirements.txt
anthropic
openai
gitpython
dotenv
huggingface_hub